+++ Controllers/QuestionController.php
@@ -23,19 +23,20 @@ class QuestionsController
         echo json_encode(['data' => $result]);
     }
     // lấy ảnh của câu hỏi
-    public function getImageAnswer($id){
+    public function getImageAnswer($id)
+    {
         $result = $this->QuestionModel->getImageAnswerModel($id);
         echo json_encode(['data' => $result]);
     }
     public function create()
     {
-        $data2 = json_decode(file_get_contents("php://input"), true);
-        $this->QuestionModel->create($data2);
-        // if ($this->QuestionModel->create($data) == false) {
-        //     echo json_encode(['message' => "Có lỗi xảy ra !"]);
-        // } else {
-        //     echo json_encode(['message' => "Tạo mới bài thi thành công !"]);
-        // }
+        $checkToken = CheckToken::checkToken();
+        if ($checkToken) {
+            $data2 = json_decode(file_get_contents("php://input"), true);
+            $this->QuestionModel->create($data2);
+        } else {
+            echo json_encode(['message' => "Token không hợp lệ"]);
+        }
     }
     public function getExam($id)
     {
@@ -52,34 +53,48 @@ class QuestionsController
     }
     public function update($id)
     {
-        $data2 = json_decode(file_get_contents("php://input"), true);
-        if ($id == 0) {
-            echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
-        } else {
-            // $this->QuestionModel->update($data2, $id);
-            if ($this->QuestionModel->update($data2, $id) == false) {
-                echo json_encode(['message' => 'Cập nhật câu hỏi không thành công !']);
-            } else {
-                echo json_encode(['message' => 'Cập nhật thông tin câu hỏi thành công !']);
+        $checkToken = CheckToken::checkToken();
+        if ($checkToken === true) {
+            $data2 = json_decode(file_get_contents("php://input"), true);
+            // kiểm tra dữ liệu tránh truyền script vào input
+            foreach ($data2 as $key => $value) {
+                $data2[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
             }
+            if ($id == 0) {
+                echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
+            } else {
+                // $this->QuestionModel->update($data2, $id);
+                if ($this->QuestionModel->update($data2, $id) == false) {
+                    echo json_encode(['message' => 'Cập nhật câu hỏi không thành công !']);
+                } else {
+                    echo json_encode(['message' => 'Cập nhật thông tin câu hỏi thành công !']);
+                }
+            }
+        } else {
+            echo json_encode(['message' => "Token không hợp lệ"]);
         }
     }
 
     public function delete($id)
     {
-        if ($id == 0) {
-            echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
-        } else {
-            // $this->QuestionModel->delete($id);
-            if ($this->QuestionModel->delete($id) == false) {
-                echo json_encode(['message' => 'Có lỗi xảy ra !']);
+        $checkToken = CheckToken::checkToken();
+        if ($checkToken) {
+            if ($id == 0) {
+                echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
             } else {
-                echo json_encode(['message' => 'Xóa bài thi thành công !']);
+                // $this->QuestionModel->delete($id);
+                if ($this->QuestionModel->delete($id) == false) {
+                    echo json_encode(['message' => 'Có lỗi xảy ra !']);
+                } else {
+                    echo json_encode(['message' => 'Xóa bài thi thành công !']);
+                }
             }
+        } else {
+            echo json_encode(['message' => "Token không hợp lệ"]);
         }
     }
     public function getUser()
     {
-       return $this->QuestionModel->getUserCreate();
+        return $this->QuestionModel->getUserCreate();
     }
 }
