+++ Controllers/ResultController.php
@@ -18,7 +18,8 @@ class ResultController
         echo json_encode(['data' => $result]);
     }
     // lấy danh sách các bài làm của user
-    public function getResultListUser($id){
+    public function getResultListUser($id)
+    {
         $this->ResultModel->getUserResultListModel($id);
     }
     public function detail($id)
@@ -26,39 +27,64 @@ class ResultController
         $result = $this->ResultModel->read($id);
         echo json_encode(['data' => $result]);
     }
-    public function getReview($id){
+    public function getReview($id)
+    {
         $this->ResultModel->getReviewModel($id);
     }
     public function create()
     {
-        $data = json_decode(file_get_contents("php://input"), true);
-        $this->ResultModel->createResultExam($data);
+        $checkToken = CheckToken::checkToken();
+        if ($checkToken === true) {
+            $data = json_decode(file_get_contents("php://input"), true);
+            // kiểm tra dữ liệu tránh truyền script vào input
+            foreach ($data as $key => $value) {
+                $data[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
+            }
+            $this->ResultModel->createResultExam($data);
+        } else {
+            echo json_encode(['message' => "Token không hợp lệ"]);
         }
-    
+    }
+
     public function update($id)
     {
-        $data = json_decode(file_get_contents("php://input"), true);
-        if ($id == 0) {
-            echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
-        } else {
-            if ($this->ResultModel->update($data, $id) == false) {
-                echo json_encode(['message' => 'Cập nhật bài thi không thành công !']);
-            } else {
-                echo json_encode(['message' => 'Cập nhật thông tin bài thi thành công !']);
+        $checkToken = CheckToken::checkToken();
+        if ($checkToken === true) {
+            $data = json_decode(file_get_contents("php://input"), true);
+            // kiểm tra dữ liệu tránh truyền script vào input
+            foreach ($data as $key => $value) {
+                $data[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
             }
+            if ($id == 0) {
+                echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
+            } else {
+                if ($this->ResultModel->update($data, $id) == false) {
+                    echo json_encode(['message' => 'Cập nhật bài thi không thành công !']);
+                } else {
+                    echo json_encode(['message' => 'Cập nhật thông tin bài thi thành công !']);
+                }
+            }
+        } else {
+            echo json_encode(['message' => "Token không hợp lệ"]);
         }
     }
 
     public function delete($id)
     {
-        if ($id == 0) {
-            echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
-        } else {
-            if ($this->ResultModel->delete($id) == false) {
-                echo json_encode(['message' => 'Có lỗi xảy ra !']);
+
+        $checkToken = CheckToken::checkToken();
+        if ($checkToken === true) {
+            if ($id == 0) {
+                echo json_encode(['message' => 'Dữ liệu bài thi không tồn tại !']);
             } else {
-                echo json_encode(['message' => 'Xóa bài thi thành công !']);
+                if ($this->ResultModel->delete($id) == false) {
+                    echo json_encode(['message' => 'Có lỗi xảy ra !']);
+                } else {
+                    echo json_encode(['message' => 'Xóa bài thi thành công !']);
+                }
             }
+        } else {
+            echo json_encode(['message' => "Token không hợp lệ"]);
         }
     }
 }
